+++ packages/remix-dev/compiler/watch.ts
@@ -2,9 +2,11 @@ import chokidar from "chokidar";
 import debounce from "lodash.debounce";
 import * as path from "path";
 
-import { type RemixConfig, readConfig } from "../config";
+import type { RemixConfig } from "../config";
+import { readConfig } from "../config";
+import type { AssetsManifest } from "./assets";
 import { logCompileFailure } from "./onCompileFailure";
-import { type CompileOptions } from "./options";
+import type { CompileOptions } from "./options";
 import { compile, createRemixCompiler, dispose } from "./remixCompiler";
 import { warnOnce } from "./warnings";
 
@@ -19,8 +21,9 @@ function isEntryPoint(config: RemixConfig, file: string): boolean {
 }
 
 export type WatchOptions = Partial<CompileOptions> & {
+  reloadConfig?(root: string): Promise<RemixConfig>;
   onRebuildStart?(): void;
-  onRebuildFinish?(durationMs: number): void;
+  onRebuildFinish?(durationMs: number, assetsManifest?: AssetsManifest): void;
   onFileCreated?(file: string): void;
   onFileChanged?(file: string): void;
   onFileDeleted?(file: string): void;
@@ -31,8 +34,10 @@ export async function watch(
   config: RemixConfig,
   {
     mode = "development",
+    liveReloadPort,
     target = "node14",
     sourcemap = true,
+    reloadConfig = readConfig,
     onWarning = warnOnce,
     onCompileFailure = logCompileFailure,
     onRebuildStart,
@@ -45,6 +50,7 @@ export async function watch(
 ): Promise<() => Promise<void>> {
   let options: CompileOptions = {
     mode,
+    liveReloadPort,
     target,
     sourcemap,
     onCompileFailure,
@@ -64,22 +70,22 @@ export async function watch(
     dispose(compiler);
 
     try {
-      config = await readConfig(config.rootDirectory);
+      config = await reloadConfig(config.rootDirectory);
     } catch (error: unknown) {
       onCompileFailure(error as Error);
       return;
     }
 
     compiler = createRemixCompiler(config, options);
-    await compile(compiler);
-    onRebuildFinish?.(Date.now() - start);
+    let assetsManifest = await compile(compiler);
+    onRebuildFinish?.(Date.now() - start, assetsManifest);
   }, 500);
 
   let rebuild = debounce(async () => {
     onRebuildStart?.();
     let start = Date.now();
-    await compile(compiler, { onCompileFailure });
-    onRebuildFinish?.(Date.now() - start);
+    let assetsManifest = await compile(compiler, { onCompileFailure });
+    onRebuildFinish?.(Date.now() - start, assetsManifest);
   }, 100);
 
   let toWatch = [config.appDirectory];
@@ -109,7 +115,7 @@ export async function watch(
       onFileCreated?.(file);
 
       try {
-        config = await readConfig(config.rootDirectory);
+        config = await reloadConfig(config.rootDirectory);
       } catch (error: unknown) {
         onCompileFailure(error as Error);
         return;